+++ folio-kafka-wrapper/src/main/java/org/folio/kafka/KafkaConfig.java
@@ -4,10 +4,13 @@ import io.kcache.KafkaCacheConfig;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
+import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.producer.ProducerConfig;
+import org.apache.kafka.common.config.SslConfigs;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 
@@ -32,6 +35,28 @@ public class KafkaConfig {
   public static final String KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG = "kafka.consumer.max.poll.interval.ms";
   public static final String KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG_DEFAULT = "300000";
 
+  public static final String KAFKA_SECURITY_PROTOCOL_CONFIG = "security.protocol";
+  public static final String KAFKA_SECURITY_PROTOCOL_DEFAULT = "PLAINTEXT";
+
+  public static final String KAFKA_SSL_PROTOCOL_CONFIG = "ssl.protocol";
+  public static final String KAFKA_SSL_PROTOCOL_DEFAULT = "TLSv1.2";
+
+  public static final String KAFKA_SSL_KEY_PASSWORD_CONFIG = "ssl.key.password";
+
+  public static final String KAFKA_SSL_TRUSTSTORE_LOCATION_CONFIG = "ssl.truststore.location";
+
+  public static final String KAFKA_SSL_TRUSTSTORE_PASSWORD_CONFIG = "ssl.truststore.password";
+
+  public static final String KAFKA_SSL_TRUSTSTORE_TYPE_CONFIG = "ssl.truststore.type";
+  public static final String KAFKA_SSL_TRUSTSTORE_TYPE_DEFAULT = "JKS";
+
+  public static final String KAFKA_SSL_KEYSTORE_LOCATION_CONFIG = "ssl.keystore.location";
+
+  public static final String KAFKA_SSL_KEYSTORE_PASSWORD_CONFIG = "ssl.keystore.password";
+
+  public static final String KAFKA_SSL_KEYSTORE_TYPE_CONFIG = "ssl.keystore.type";
+  public static final String KAFKA_SSL_KEYSTORE_TYPE_DEFAULT = "JKS";
+
   private static final String KAFKA_CACHE_TOPIC_PROPERTY = "kafkacache.topic";
   private static final String KAFKA_CACHE_TOPIC_PROPERTY_DEFAULT = "events_cache";
 
@@ -47,6 +72,7 @@ public class KafkaConfig {
     producerProps.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, "true");
     producerProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
     producerProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
+    ensureSecurityProps(producerProps);
     return producerProps;
   }
 
@@ -54,12 +80,18 @@ public class KafkaConfig {
     Map<String, String> consumerProps = new HashMap<>();
     consumerProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, getKafkaUrl());
     consumerProps.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "false");
-    consumerProps.put(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, SimpleConfigurationReader.getValue(KAFKA_CONSUMER_MAX_POLL_RECORDS_CONFIG, KAFKA_CONSUMER_MAX_POLL_RECORDS_CONFIG_DEFAULT));
-    consumerProps.put(ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG, SimpleConfigurationReader.getValue(KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG, KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG_DEFAULT));
-    consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, SimpleConfigurationReader.getValue(KAFKA_CONSUMER_AUTO_OFFSET_RESET_CONFIG, KAFKA_CONSUMER_AUTO_OFFSET_RESET_CONFIG_DEFAULT));
-    consumerProps.put(ConsumerConfig.METADATA_MAX_AGE_CONFIG, SimpleConfigurationReader.getValue(KAFKA_CONSUMER_METADATA_MAX_AGE_CONFIG, KAFKA_CONSUMER_METADATA_MAX_AGE_CONFIG_DEFAULT));
     consumerProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer");
     consumerProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer");
+
+    consumerProps.put(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_CONSUMER_MAX_POLL_RECORDS_CONFIG, SpringKafkaProperties.KAFKA_CONSUMER_MAX_POLL_RECORDS), KAFKA_CONSUMER_MAX_POLL_RECORDS_CONFIG_DEFAULT));
+    consumerProps.put(ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG, SpringKafkaProperties.KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS), KAFKA_CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG_DEFAULT));
+    consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_CONSUMER_AUTO_OFFSET_RESET_CONFIG, SpringKafkaProperties.KAFKA_CONSUMER_AUTO_OFFSET_RESET), KAFKA_CONSUMER_AUTO_OFFSET_RESET_CONFIG_DEFAULT));
+    consumerProps.put(ConsumerConfig.METADATA_MAX_AGE_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_CONSUMER_METADATA_MAX_AGE_CONFIG, SpringKafkaProperties.KAFKA_CONSUMER_METADATA_MAX_AGE), KAFKA_CONSUMER_METADATA_MAX_AGE_CONFIG_DEFAULT));
+    ensureSecurityProps(consumerProps);
     return consumerProps;
   }
 
@@ -77,6 +109,27 @@ public class KafkaConfig {
 
   public int getNumberOfPartitions() {
     return Integer.parseInt(SimpleConfigurationReader.getValue(KAFKA_NUMBER_OF_PARTITIONS, KAFKA_NUMBER_OF_PARTITIONS_DEFAULT));
+  }
+
+  private void ensureSecurityProps(Map<String, String> clientProps) {
+    clientProps.put(CommonClientConfigs.SECURITY_PROTOCOL_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SECURITY_PROTOCOL_CONFIG, SpringKafkaProperties.KAFKA_SECURITY_PROTOCOL), KAFKA_SECURITY_PROTOCOL_DEFAULT));
+    clientProps.put(SslConfigs.SSL_PROTOCOL_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_PROTOCOL_CONFIG, SpringKafkaProperties.KAFKA_SSL_PROTOCOL), KAFKA_SSL_PROTOCOL_DEFAULT));
+    clientProps.put(SslConfigs.SSL_KEY_PASSWORD_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_KEY_PASSWORD_CONFIG, SpringKafkaProperties.KAFKA_SSL_KEY_PASSWORD), null));
+    clientProps.put(SslConfigs.SSL_TRUSTSTORE_LOCATION_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_TRUSTSTORE_LOCATION_CONFIG, SpringKafkaProperties.KAFKA_SSL_TRUSTSTORE_LOCATION), null));
+    clientProps.put(SslConfigs.SSL_TRUSTSTORE_PASSWORD_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_TRUSTSTORE_PASSWORD_CONFIG, SpringKafkaProperties.KAFKA_SSL_TRUSTSTORE_PASSWORD), null));
+    clientProps.put(SslConfigs.SSL_TRUSTSTORE_TYPE_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_TRUSTSTORE_TYPE_CONFIG, SpringKafkaProperties.KAFKA_SSL_TRUSTSTORE_TYPE), KAFKA_SSL_TRUSTSTORE_TYPE_DEFAULT));
+    clientProps.put(SslConfigs.SSL_KEYSTORE_LOCATION_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_KEYSTORE_LOCATION_CONFIG, SpringKafkaProperties.KAFKA_SSL_KEYSTORE_LOCATION), null));
+    clientProps.put(SslConfigs.SSL_KEYSTORE_PASSWORD_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_KEYSTORE_PASSWORD_CONFIG, SpringKafkaProperties.KAFKA_SSL_KEYSTORE_PASSWORD), null));
+    clientProps.put(SslConfigs.SSL_KEYSTORE_TYPE_CONFIG, SimpleConfigurationReader.getValue(
+      List.of(KAFKA_SSL_KEYSTORE_TYPE_CONFIG, SpringKafkaProperties.KAFKA_SSL_KEYSTORE_TYPE), KAFKA_SSL_KEYSTORE_TYPE_DEFAULT));
   }
 
 }